feat: restructure dashboard and overview components with new page layout and stats card implementation
