+++ src/features/dashboard/core/comps/dashboard-page-title.tsx
@@ -1,7 +1,7 @@
 type DashboardTitleProps = {
-    title: string;
+    children: string;
 };
 
-export const DashboardPageTitle = ({ title }: DashboardTitleProps) => {
-    return <span className="font-satoshi font-medium text-2xl leading-[34px] text-stone-950">{title}</span>;
+export const DashboardPageTitle = ({ children }: DashboardTitleProps) => {
+    return <span className="font-satoshi font-medium text-2xl leading-[34px] text-stone-950">{children}</span>;
 };
